seperated myRanges.hpp to different files
